requirements.txt
# Required Python packages get listed here, one per line.

runpod~=1.7.9
llama-cpp-python==0.2.90
huggingface-hub==0.25.2
pydantic==2.9.2
